Inspired by a conversation with TristAHN, changed the flow of pages for browse and cluster queries to be more REST-ful.  a query "object" is cached when a POST to browse/cluster is done.  Then the query is run and cached (if not already cached) when a GET to orth_query is done.  Then a GET is done to orth_result to display the result.
Previously a browse POST would kick off the running of the query and then a GET to orth_result would display the result.  Wait, I know that sounds simpler, but you have to see the code!  The logic of whether a query is cached or needs to be run synchronously or asynchronously has been centralized in orth_query, while the logic of creating a browse query with browse ids and seq ids is in the browse page.

Well... anyway...