readme.uniprot.txt has commands and notes describing uniprot as a data source for complete genomes.
I am mostly through the process of loading a dataset into the database.  Some more metadata will have to be added.
The roundup_db code has also been changed to use a "version", so there can be multiple datasets in the database,
and so a dataset can be loaded (which might take days) and then the website can be switched over without any
disruption of the site or inconsistent state during those loading days.

There is some more data (about genomes and sources) that needs to be put into the website or db.

The web code needs to be modified to use some of these new sources of info.

But I am close to getting datasets, versions, and uniprot working!