Changes to lsf.py and lsfdispatch.py to make them simpler, more robust, and less quirky.
added path components.  now you can dispatch to a function not in the same dir as lsfdispatch.
adde function arguments.  now not all parameters need to be specified as keywords.  This is more natural.
can save serialized function output to a file now.  previously the output was sent to stdout.
redirecting lsf job output to /dev/null is now optional, and on by default.  this makes it easier to redirect job output to a file or get a job report email.

changed lsf.submitToLSF to lsf.bsub
removed lsf.py environment setup, which was orchestra specific.  replaced with a func to configure the environment which is called by config.py