Add speed and taxonomy selection to genome selection
Big ups to Byron for teaming up with me to create a combination of
checkboxes for filtering by taxonomy, a select box for choosing
genomes, and a textarea for pasting in the genomes.  This 3-headed
beast replaced the chosen multiselect.  Not entirely sure how I feel
about that, but it is fast.  And I like the idea that you can paste.

Moved javascript into its own file.