Avoid persistent db connections.  Refactor Stats and Dones.
In roundup.dataset, do not keep an open connection associated with Stats()
objects, since scaled up to 1000+ processes, mysql starts to throw 'Too
many connections' errors.

I think this is the only persistently open connection used during
dataset computation.

Added a close() method to dbutil.Reuser, so the connection can be
closed.  This allows dataset.extractPerformanceStats to open a
connection, get a bunch of stats rapidly, and then close the connection,
via dataset.statsCM().

Tweaked whitespace in kvstore.py.
Added some docs and an assertion to orchmysql.

dataset.Stats() has two usage patterns:

- Many processes occasionally put some stats.
- One process rapidly gets many stats.

To accommodate this pattern, I need to be able to have a non-persistent
connection (open, do something, close) for putting stats and a
semi-persistent connection for getting stats (open, do many things,
close).  This is accomplished by creating a context manager for a stats
object, which closes the connection when exiting the context.  That is
statsCM().  For putting stats one uses getStats(), which creates a
kvstore with a ClosingFactoryCM.

While implementing this stats contextmanager, I realized that it makes
more sense (to me) to inject Stats (and Dones) with a KVStore (and
KStore) instead of an open_conn() factory function.  This way they
depend on what they use and are not coupled to kvstore object creation
or dbapi connections unnecessarily.