Fix rollback() call on NoneType connection object.
This patch fixes the source of the problem, orchmysql.openConn, which
incorrectly failed to reraise an exception when it was done retrying to
establish a connection.  The patch also fixes the proximal cause by
checking that conn is not None before calling rollback on it.  This will
avoid having the rollback exception mask the exception it is trying to
partially recover from.