Fix orthology query lsf queue and roundup_db.py
Internally to roundup_db.py, I forgot to pass release as a parameter for
some function calls.  Fixed.

The 'short' lsf queue roundup_util.py uses to run ortholog is often very
backed up.  As in hours to run a query instead of seconds.  Change to
cbi_12h.  It should be less busy, even though it is has only a few job
slots.