Rename dburl to open_url_cm. Fix bug w/ factory cm
Ugh.  Some code, like roundup_db, uses a function that creates a context
manager (cm) that opens and yields a connection.  Other code, like kvstore, uses
a context manager that opens and yields a connection repeatedly.
Passing the function to kvstore does not work b/c kvstore does not call
the function to get a context manager.  It simply uses the context
manager it is given to get a connection.