Use a CLI for dataset computation, not workflow.
In order to make dataset.py more command-line-centric, the dispatch of
function calls onto lsf by workflow.py was replaced with the following
pieces in dataset:

- Creation of a command line interface for computing a job. I.e. `python
  dataset.py compute_job <dataset> <job>`
- Tracking of running jobs and done jobs in dataset.py
- Composing command lines and job names for use with dones and lsf.
- Submission of jobs to lsf.

The removal of workflow.py meant removing some of the parallelization of
other dataset steps, like formatting genomes.  However this
parallelization was not necessary and now the code is simpler and more
legible.