Do not get db creds from env.
For some reason, env vars are defined for some
db creds, which was forcing the dev roundup website
to use the devroundup database, even when I wanted
to test against the prod (roundup) database.  The
fix is to just use the creds in secrets.py, and
ignore the environment.
This is OK, since I do not use the environment to configure
creds currently, either for the website or dataset computation.